backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
pydantic[email]==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-docx==1.1.2
pdfplumber==0.11.6
PyMuPDF==1.25.5
Pillow==11.2.1
httpx==0.28.1
reportlab==4.4.0
email-validator==2.2.0
